requirements.txt
keras==3.9.2
numpy==2.2.6
opencv_python_headless==4.11.0.86
protobuf==6.31.1
streamlit==1.45.1
google-generativeai==0.8.5
